+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## 3.1.0 - TBD
+### Added
+- Custom source selection
+
 ### Fixed
 - Fixed division by zero when not copying any new files
 - Fixed crash on CLI mode when loading destination drives